bookTable

+++ madangTest/src/main/webapp/deleteBookProcess.jsp
@@ -9,13 +9,34 @@
 <jsp:useBean id="dao" class ="com.sist.dao.BookDAO"/>
 <body>
 	<%
-	int bookid = Integer.parseInt(request.getParameter("bookid"));
-	int re = dao.deleteBook(bookid);
-	
-	if(re > 0)
-		response.sendRedirect("bookList.jsp");
-	else
-		out.print("삭제실패");
+	if(request.getParameter("bookid")!=null){
+		int bookid = Integer.parseInt(request.getParameter("bookid"));
+		int re = dao.deleteBook(bookid);
+		
+		if(re > 0){
+			%>
+			<script type="text/javascript">
+				alert("삭제 확인");
+				location.href = "bookList.jsp";
+			</script>	
+			<%
+		}
+		else{
+		%>
+			<script type="text/javascript">
+				alert("삭제 실패");
+			</script>
+		<%
+		}
+	}
+	else{
+		%>
+		<script type="text/javascript">
+			alert("삭제할 도서번호 확인");
+			location.href("bookList.jsp");
+		</script>
+		<%	
+	}
 	%>
 </body>
 </html>
